backend/pyproject.toml
[project]
name = "pm-backend"
version = "0.1.0"
description = "FastAPI backend for PM MVP"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: calls real OpenRouter (requires OPENROUTER_API_KEY and network)",
]
